u32 framebuffer size

+++ source/client/renderer/renderer.cpp
@@ -313,7 +313,7 @@ static void CreateRenderTargets( bool first_time ) {
 	}, first_time ? NONE : Optional( frame_static.render_targets.shadowmap ) );
 }
 
-void RendererBeginFrame( SDL_Window * window, int viewport_width, int viewport_height, bool minimized, bool fullscreen_exclusive ) {
+void RendererBeginFrame( SDL_Window * window, u32 viewport_width, u32 viewport_height, bool minimized, bool fullscreen_exclusive ) {
 	TracyZoneScoped;
 
 	HotloadShaders();
@@ -326,11 +326,11 @@ void RendererBeginFrame( SDL_Window * window, int viewport_width, int viewport_h
 
 	RenderBackendWaitForNewFrame();
 
-	frame_static.viewport_width = Max2( 1, viewport_width );
-	frame_static.viewport_height = Max2( 1, viewport_height ),
+	frame_static.viewport_width = Max2( 1_u32, viewport_width );
+	frame_static.viewport_height = Max2( 1_u32, viewport_height ),
 	frame_static.viewport = Vec2( viewport_width, viewport_height );
 	frame_static.minimized = minimized;
-	frame_static.viewport_resized = viewport_width > 0 && viewport_height > 0 && ( u32( viewport_width ) != last_viewport_width || u32( viewport_height ) != last_viewport_height );
+	frame_static.viewport_resized = viewport_width > 0 && viewport_height > 0 && ( viewport_width != last_viewport_width || viewport_height != last_viewport_height );
 	frame_static.aspect_ratio = float( frame_static.viewport_width ) / float( frame_static.viewport_height );
 	frame_static.msaa_samples = r_samples->integer;
 	frame_static.shadow_quality = ShadowQuality( r_shadow_quality->integer );
